+++ identity-authentication/source/uim/platform/identity/authentication/domain/ports/repositories/tenant.d
@@ -15,7 +15,7 @@ mixin(ShowModule!());
 interface TenantRepository {
   Tenant findById(TenantId id);
   Tenant findBySubdomain(string subdomain);
-  Tenant[] findAll(uint offset = 0, uint limit = 100);
+  Tenant[] findAll(size_t offset = 0, size_t limit = 100);
   void save(Tenant tenant);
   void update(Tenant tenant);
   void remove(TenantId id);
