+++ identity-directory/source/uim/platform/identity/directory/application/usecases/manage/schemas.d
@@ -54,7 +54,7 @@ class ManageSchemasUseCase { // TODO: UIMUseCase {
   }
 
   /// List schemas for a tenant.
-  Schema[] listSchemas(TenantId tenantId, uint offset = 0, uint limit = 100) {
+  Schema[] listSchemas(TenantId tenantId, size_t offset = 0, size_t limit = 100) {
     return schemaRepo.findByTenant(tenantId, offset, limit);
   }
