+++ identity-directory/source/uim/platform/identity/directory/application/usecases/query_audit_log.d
@@ -22,29 +22,29 @@ class QueryAuditLogUseCase { // TODO: UIMUseCase {
   }
 
   /// List audit events by tenant.
-  AuditEvent[] listEvents(TenantId tenantId, uint offset = 0, uint limit = 100) {
+  AuditEvent[] listEvents(TenantId tenantId, size_t offset = 0, size_t limit = 100) {
     return auditRepo.findByTenant(tenantId, offset, limit);
   }
 
   /// Find events by actor.
-  AuditEvent[] findByActor(string actorId, uint offset = 0, uint limit = 100) {
+  AuditEvent[] findByActor(string actorId, size_t offset = 0, size_t limit = 100) {
     return auditRepo.findByActor(actorId, offset, limit);
   }
 
   /// Find events by target resource.
-  AuditEvent[] findByTarget(string targetId, uint offset = 0, uint limit = 100) {
+  AuditEvent[] findByTarget(string targetId, size_t offset = 0, size_t limit = 100) {
     return auditRepo.findByTarget(targetId, offset, limit);
   }
 
   /// Find events by type.
   AuditEvent[] findByType(TenantId tenantId, AuditEventType eventType,
-      uint offset = 0, uint limit = 100) {
+      size_t offset = 0, size_t limit = 100) {
     return auditRepo.findByType(tenantId, eventType, offset, limit);
   }
 
   /// Find events within a time range.
   AuditEvent[] findByTimeRange(TenantId tenantId, long from, long to,
-      uint offset = 0, uint limit = 100) {
+      size_t offset = 0, size_t limit = 100) {
     return auditRepo.findByTimeRange(tenantId, from, to, offset, limit);
   }
 }
